pyproject.toml
[project]
name = "claude-api"
version = "0.1.0"
description = "Learning Claude API"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "anthropic>=0.76.0",
    "python-dotenv>=1.2.1",
]
